Fixed bug #799 and converted legend measurements to floating-point. Also fixed a bug where the legend lines would extend past their endpoints.
git-svn-id: http://svn.osgeo.org/qgis/trunk/qgis@7638 c8812cc2-4d05-0410-92ff-de0c093fc19c

+++ src/app/composer/qgscomposervectorlegend.h
@@ -185,10 +185,10 @@ private:
     // Pen
     QPen  mPen;
 
-    int mMargin;
-    int mSymbolHeight;
-    int mSymbolWidth;
-    int mSymbolSpace;
+    double mMargin;
+    double mSymbolHeight;
+    double mSymbolWidth;
+    double mSymbolSpace;
 
     // Cache used in composer preview
     // NOTE:  QCanvasView is slow with bigger images but the spped does not decrease with image size.
